refactor: Move dialogs to feature-specific packages (#203)
- Moved dialog composables from `common.component.dialog` to feature-specific packages under `presentation`
  - `FullImageDialog` & `RequestAcceptDialog` → `presentation.rentaldetail.dialog`
  - `FullImagePagerDialog` → `presentation.productdetail.dialog`
  - `NetworkErrorDialog`, `ServerErrorDialog`, `SessionExpiredDialog` → `presentation.main.dialog`
- Updated import statements in relevant screens/routes
- Simplified `RequestAcceptDialog` by passing parameters directly instead of using `RequestAcceptDialogUiModel`
