+++ app/src/main/java/com/example/rentit/presentation/main/dialog/NetworkErrorDialog.kt
@@ -1,8 +1,9 @@
-package com.example.rentit.common.component.dialog
+package com.example.rentit.presentation.main.dialog
 
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.res.stringResource
 import com.example.rentit.R
+import com.example.rentit.common.component.dialog.BaseDialog
 
 @Composable
 fun NetworkErrorDialog(navigateBack: () -> Unit, onRetry: () -> Unit) {
