+++ app/src/main/java/com/example/rentit/presentation/productdetail/dialog/FullImagePagerDialog.kt
@@ -1,4 +1,4 @@
-package com.example.rentit.common.component.dialog
+package com.example.rentit.presentation.productdetail.dialog
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.clickable
@@ -6,6 +6,9 @@ import androidx.compose.foundation.interaction.MutableInteractionSource
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.fillMaxHeight
 import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.pager.HorizontalPager
+import androidx.compose.foundation.pager.PagerState
+import androidx.compose.foundation.pager.rememberPagerState
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.remember
 import androidx.compose.ui.Alignment
@@ -20,8 +23,9 @@ import com.example.rentit.common.theme.AppBlack
 import com.example.rentit.common.theme.RentItTheme
 
 @Composable
-fun FullImageDialog(
-    imageUrls: String?,
+fun FullImagePagerDialog(
+    pagerState: PagerState,
+    imageUrls: List<String?>,
     onDismiss: () -> Unit,
 ) {
     Dialog(
@@ -41,23 +45,34 @@ fun FullImageDialog(
                 ) { onDismiss() },
             contentAlignment = Alignment.Center
         ) {
-            LoadableUrlImage(
-                modifier = Modifier.fillMaxHeight(0.7f),
-                imgUrl = imageUrls,
-                defaultImageResId = R.drawable.img_placeholder,
-                defaultDescResId = R.string.screen_product_detail_img_description,
-                contentScale = ContentScale.FillWidth
-            )
+            HorizontalPager(
+                state = pagerState,
+                verticalAlignment = Alignment.CenterVertically
+            ) { page ->
+                LoadableUrlImage(
+                    modifier = Modifier.fillMaxHeight(0.7f),
+                    imgUrl = imageUrls[page],
+                    defaultImageResId = R.drawable.img_placeholder,
+                    defaultDescResId = R.string.screen_product_detail_img_description,
+                    contentScale = ContentScale.FillWidth
+                )
+            }
         }
     }
 }
 
 @Preview(showBackground = true)
 @Composable
-fun FullImageDialogPreview() {
+fun FullImagePagerDialogPreview() {
+    val sampleImageUrls =
+        listOf(
+            "https://www.econovill.com/news/photo/201812/353683_236958_5647.jpg",
+            "https://flexible.img.hani.co.kr/flexible/normal/590/590/imgdb/resize/2007/1227/68227042_20071227.jpg"
+            )
     RentItTheme {
-        FullImageDialog(
-            imageUrls = "sampleImageUrls",
+        FullImagePagerDialog(
+            pagerState = rememberPagerState(1) { sampleImageUrls.size },
+            imageUrls = sampleImageUrls,
             onDismiss = {},
         )
     }
