+++ app/src/main/java/com/example/rentit/presentation/rentaldetail/RentalDetailRoute.kt
@@ -21,7 +21,7 @@ import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import androidx.lifecycle.repeatOnLifecycle
 import androidx.navigation.NavHostController
 import com.example.rentit.R
-import com.example.rentit.common.component.dialog.RequestAcceptDialog
+import com.example.rentit.presentation.rentaldetail.dialog.RequestAcceptDialog
 import com.example.rentit.common.component.layout.LoadingScreen
 import com.example.rentit.common.enums.RentalRole
 import com.example.rentit.navigation.chatroom.navigateToChatRoom
