+++ app/src/main/java/com/example/rentit/presentation/rentaldetail/dialog/FullImageDialog.kt
@@ -1,4 +1,4 @@
-package com.example.rentit.common.component.dialog
+package com.example.rentit.presentation.rentaldetail.dialog
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.clickable
@@ -6,9 +6,6 @@ import androidx.compose.foundation.interaction.MutableInteractionSource
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.fillMaxHeight
 import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.pager.HorizontalPager
-import androidx.compose.foundation.pager.PagerState
-import androidx.compose.foundation.pager.rememberPagerState
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.remember
 import androidx.compose.ui.Alignment
@@ -23,9 +20,8 @@ import com.example.rentit.common.theme.AppBlack
 import com.example.rentit.common.theme.RentItTheme
 
 @Composable
-fun FullImagePagerDialog(
-    pagerState: PagerState,
-    imageUrls: List<String?>,
+fun FullImageDialog(
+    imageUrls: String?,
     onDismiss: () -> Unit,
 ) {
     Dialog(
@@ -45,34 +41,23 @@ fun FullImagePagerDialog(
                 ) { onDismiss() },
             contentAlignment = Alignment.Center
         ) {
-            HorizontalPager(
-                state = pagerState,
-                verticalAlignment = Alignment.CenterVertically
-            ) { page ->
-                LoadableUrlImage(
-                    modifier = Modifier.fillMaxHeight(0.7f),
-                    imgUrl = imageUrls[page],
-                    defaultImageResId = R.drawable.img_placeholder,
-                    defaultDescResId = R.string.screen_product_detail_img_description,
-                    contentScale = ContentScale.FillWidth
-                )
-            }
+            LoadableUrlImage(
+                modifier = Modifier.fillMaxHeight(0.7f),
+                imgUrl = imageUrls,
+                defaultImageResId = R.drawable.img_placeholder,
+                defaultDescResId = R.string.screen_product_detail_img_description,
+                contentScale = ContentScale.FillWidth
+            )
         }
     }
 }
 
 @Preview(showBackground = true)
 @Composable
-fun FullImagePagerDialogPreview() {
-    val sampleImageUrls =
-        listOf(
-            "https://www.econovill.com/news/photo/201812/353683_236958_5647.jpg",
-            "https://flexible.img.hani.co.kr/flexible/normal/590/590/imgdb/resize/2007/1227/68227042_20071227.jpg"
-            )
+fun FullImageDialogPreview() {
     RentItTheme {
-        FullImagePagerDialog(
-            pagerState = rememberPagerState(1) { sampleImageUrls.size },
-            imageUrls = sampleImageUrls,
+        FullImageDialog(
+            imageUrls = "sampleImageUrls",
             onDismiss = {},
         )
     }
